Upgrade PyCDS to modern version and rewrite tests
PyCDS has been dreadfully behind. Unfortunately, in the many updated
versions since we have upgraded it, we completely changed the
testing data and testing session methodology. This required extensive
hacking on the pdp_util code. For the record, the pdp_util code's lack
of session manangement is horrible and the way that I had to shim
sessions into it is not one of my better moments. But this will allow us
to move forward and potentially rewrite this whole package to me much
simpler at a later date. Really, the package doesn't add a lot of value
anymore, and we should probably just rewrite it with a couple of Flask
methods directly into pdp. But... for another day.